Fix issuses with systemd and mysql server packages on openSUSE.
On openSUSE init scripts are handled by systemd. Unfortunatley mysql server
packages (both mariadb and mysql-community-server) are not executing proper
actions to reload systemd after installing new init scripts.

Fix was allready send to openSUSE packages but this recipie can be ran on
unpatched system, so we need to make sure that system is reloaded before
starting mysql service.

Fix also is aware of problems with lsb on SLE that's why it checks
if node["lsb"]["description"] exists.